notes
Suffice it to say, this is not nearly as flexible a dialog styling as I was hoping for. Setting an explicit width for the dialog (in CSS no less) offers no flexibility for the developer to make a practically full-screen dialog, vs. a message popup sort of dialog. The question you have to ask yourself in styling a dialog is: how do you, with pure CSS, style a div or whatever to be vertically and horizontally centered, while still allowing its contents to dictate its size (both width and height), and even allowing the developer to set a percentage or non-percentage width or max-width?